refactor: rename BibleItemViewController to BibleItemsViewController for consistency across components
