+++ src/AppPresenterComp.tsx
@@ -2,9 +2,9 @@ import { lazy, useMemo } from 'react';
 
 import { resizeSettingNames } from './resize-actor/flexSizeHelpers';
 import ResizeActorComp from './resize-actor/ResizeActorComp';
-import BibleItemViewController, {
+import BibleItemsViewController, {
     BibleItemViewControllerContext,
-} from './bible-reader/BibleItemViewController';
+} from './bible-reader/BibleItemsViewController';
 import SlideEditHandlerComp from './app-document-presenter/SlideEditHandlerComp';
 import BibleViewComp from './bible-reader/BibleViewComp';
 import BibleItem from './bible-list/BibleItem';
@@ -25,7 +25,7 @@ const LazyAppPresenterRightComp = lazy(() => {
 
 export default function AppPresenterComp() {
     const viewController = useMemo(() => {
-        const newViewController = new BibleItemViewController('presenter');
+        const newViewController = new BibleItemsViewController('presenter');
         newViewController.finalRenderer = (bibleItem: BibleItem) => {
             return (
                 <BibleViewTitleMaterialContext
