+++ src/bible-reader/BibleViewComp.tsx
@@ -2,7 +2,7 @@ import './BibleViewComp.scss';
 
 import BibleItem from '../bible-list/BibleItem';
 import { showAppContextMenu } from '../context-menu/AppContextMenuComp';
-import { useBibleItemViewControllerContext } from './BibleItemViewController';
+import { useBibleItemViewControllerContext } from './BibleItemsViewController';
 import {
     applyDragged,
     genDraggingClass,
